Refactor database initialization scripts
- Consolidated database initialization into separate files for better organization and maintainability.
- Removed init_db_materiali.php and integrated its functionality into init_db.php.
- Created init_db_main.php and init_db_attrezzature.php to handle main database setup and equipment table initialization respectively.
- Added error handling for missing config.php file and database connection issues.
- Streamlined table creation logic for ubicazioni, attrezzature, and logging tables.
- Improved logging and error messages for better debugging.
